# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NJAYCO OS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DivisionWindowType =
  (typeof DivisionWindowType)[keyof typeof DivisionWindowType];

export const DivisionWindowType = {
  browser: "browser",
  notepad: "notepad",
  music: "music",
  explorer: "explorer",
  admin: "admin",
  custom: "custom",
} as const;

export type DivisionStatus =
  (typeof DivisionStatus)[keyof typeof DivisionStatus];

export const DivisionStatus = {
  live: "live",
  development: "development",
  concept: "concept",
} as const;

export type DivisionCategory =
  (typeof DivisionCategory)[keyof typeof DivisionCategory];

export const DivisionCategory = {
  music: "music",
  tech: "tech",
  education: "education",
  logistics: "logistics",
  media: "media",
  corporate: "corporate",
  creative: "creative",
} as const;

export interface Division {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  fullDescription: string;
  websiteUrl?: string | null;
  iconType: string;
  iconColor: string;
  windowType: DivisionWindowType;
  notepadContent?: string | null;
  status: DivisionStatus;
  category: DivisionCategory;
  sortOrder: number;
  featured: boolean;
}

export type DivisionCreateWindowType =
  (typeof DivisionCreateWindowType)[keyof typeof DivisionCreateWindowType];

export const DivisionCreateWindowType = {
  browser: "browser",
  notepad: "notepad",
  music: "music",
  explorer: "explorer",
  admin: "admin",
  custom: "custom",
} as const;

export type DivisionCreateStatus =
  (typeof DivisionCreateStatus)[keyof typeof DivisionCreateStatus];

export const DivisionCreateStatus = {
  live: "live",
  development: "development",
  concept: "concept",
} as const;

export type DivisionCreateCategory =
  (typeof DivisionCreateCategory)[keyof typeof DivisionCreateCategory];

export const DivisionCreateCategory = {
  music: "music",
  tech: "tech",
  education: "education",
  logistics: "logistics",
  media: "media",
  corporate: "corporate",
  creative: "creative",
} as const;

export interface DivisionCreate {
  slug: string;
  name: string;
  shortDescription: string;
  fullDescription: string;
  websiteUrl?: string | null;
  iconType: string;
  iconColor: string;
  windowType: DivisionCreateWindowType;
  notepadContent?: string | null;
  status: DivisionCreateStatus;
  category: DivisionCreateCategory;
  sortOrder: number;
  featured: boolean;
}

export interface DivisionUpdate {
  id: number;
  name?: string;
  shortDescription?: string;
  websiteUrl?: string | null;
  status?: string;
  featured?: boolean;
}

export interface TrackCreate {
  artistId: number;
  albumId: number;
  title: string;
  duration: number;
  trackNumber: number;
  previewUrl?: string | null;
  featured: boolean;
}

export interface ArtistCreate {
  slug: string;
  name: string;
  bio: string;
  genre: string;
  imageUrl?: string | null;
  featured: boolean;
}

export type AlbumCreateAlbumType =
  (typeof AlbumCreateAlbumType)[keyof typeof AlbumCreateAlbumType];

export const AlbumCreateAlbumType = {
  album: "album",
  ep: "ep",
  single: "single",
} as const;

export interface AlbumCreate {
  artistId: number;
  title: string;
  coverUrl?: string | null;
  releaseYear: number;
  genre: string;
  price: number;
  albumType: AlbumCreateAlbumType;
  featured: boolean;
}

export interface Artist {
  id: number;
  slug: string;
  name: string;
  bio: string;
  genre: string;
  imageUrl?: string | null;
  featured: boolean;
}

export type AlbumAlbumType =
  (typeof AlbumAlbumType)[keyof typeof AlbumAlbumType];

export const AlbumAlbumType = {
  album: "album",
  ep: "ep",
  single: "single",
} as const;

export interface Album {
  id: number;
  artistId: number;
  artistName: string;
  title: string;
  coverUrl?: string | null;
  releaseYear: number;
  genre: string;
  price: number;
  albumType: AlbumAlbumType;
  featured: boolean;
  trackCount: number;
}

export interface Track {
  id: number;
  albumId: number;
  artistId: number;
  artistName: string;
  albumTitle: string;
  title: string;
  trackNumber: number;
  /** Duration in seconds */
  duration: number;
  audioUrl?: string | null;
  previewUrl?: string | null;
  price: number;
  isExplicit: boolean;
  streamable: boolean;
  downloadable: boolean;
  featured: boolean;
}

export type ArtistDetail = Artist & {
  albums: Album[];
  topTracks: Track[];
};

export type AlbumDetail = Album & {
  tracks: Track[];
};

export type DocumentCategory =
  (typeof DocumentCategory)[keyof typeof DocumentCategory];

export const DocumentCategory = {
  corporate: "corporate",
  investor: "investor",
  press: "press",
  legal: "legal",
  division: "division",
  services: "services",
} as const;

export type DocumentFileType =
  (typeof DocumentFileType)[keyof typeof DocumentFileType];

export const DocumentFileType = {
  txt: "txt",
  doc: "doc",
  pdf: "pdf",
  ppt: "ppt",
} as const;

export interface Document {
  id: number;
  title: string;
  filename: string;
  category: DocumentCategory;
  content: string;
  fileType: DocumentFileType;
  divisionId?: number | null;
  public: boolean;
}

export interface AdminStats {
  totalDivisions: number;
  activeDivisions: number;
  totalTracks: number;
  totalArtists: number;
  totalDocuments: number;
}

export type GetAlbumsParams = {
  artistId?: number;
};

export type GetTracksParams = {
  albumId?: number;
  artistId?: number;
};

export type GetDocumentsParams = {
  category?: string;
};