# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NJAYCO OS API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: divisions
    description: NJAYCO divisions/companies
  - name: music
    description: UV Music Group artists, albums, tracks
  - name: documents
    description: Company documents and business plans
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /divisions:
    get:
      operationId: getDivisions
      tags: [divisions]
      summary: Get all divisions
      responses:
        "200":
          description: List of divisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Division"
    post:
      operationId: createDivision
      tags: [divisions]
      summary: Create a new division (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DivisionCreate"
      responses:
        "201":
          description: Created division
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Division"
        "400":
          description: Validation error

  /divisions/{id}:
    get:
      operationId: getDivision
      tags: [divisions]
      summary: Get a division by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Division
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Division"
        "404":
          description: Not found

  /artists:
    get:
      operationId: getArtists
      tags: [music]
      summary: Get all artists
      responses:
        "200":
          description: List of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
    post:
      operationId: createArtist
      tags: [music]
      summary: Create a new artist (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistCreate"
      responses:
        "201":
          description: Created artist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Validation error

  /artists/{id}:
    get:
      operationId: getArtist
      tags: [music]
      summary: Get artist by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist with albums and tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistDetail"

  /albums:
    get:
      operationId: getAlbums
      tags: [music]
      summary: Get all albums
      parameters:
        - name: artistId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
    post:
      operationId: createAlbum
      tags: [music]
      summary: Create a new album (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumCreate"
      responses:
        "201":
          description: Created album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Validation error

  /albums/{id}:
    get:
      operationId: getAlbum
      tags: [music]
      summary: Get album by ID with tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Album with tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumDetail"

  /tracks:
    get:
      operationId: getTracks
      tags: [music]
      summary: Get tracks
      parameters:
        - name: albumId
          in: query
          required: false
          schema:
            type: integer
        - name: artistId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
    post:
      operationId: createTrack
      tags: [music]
      summary: Create a new track (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackCreate"
      responses:
        "201":
          description: Created track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Validation error

  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: Get all documents
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/divisions:
    put:
      operationId: updateDivision
      tags: [admin]
      summary: Update a division
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DivisionUpdate"
      responses:
        "200":
          description: Updated division
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Division"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Division:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        websiteUrl:
          type: string
          nullable: true
        iconType:
          type: string
        iconColor:
          type: string
        windowType:
          type: string
          enum: [browser, notepad, music, explorer, admin, custom]
        notepadContent:
          type: string
          nullable: true
        status:
          type: string
          enum: [live, development, concept]
        category:
          type: string
          enum: [music, tech, education, logistics, media, corporate, creative]
        sortOrder:
          type: integer
        featured:
          type: boolean
      required:
        - id
        - slug
        - name
        - shortDescription
        - fullDescription
        - iconType
        - iconColor
        - windowType
        - status
        - category
        - sortOrder
        - featured

    DivisionCreate:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        websiteUrl:
          type: string
          nullable: true
        iconType:
          type: string
        iconColor:
          type: string
        windowType:
          type: string
          enum: [browser, notepad, music, explorer, admin, custom]
        notepadContent:
          type: string
          nullable: true
        status:
          type: string
          enum: [live, development, concept]
        category:
          type: string
          enum: [music, tech, education, logistics, media, corporate, creative]
        sortOrder:
          type: integer
        featured:
          type: boolean
      required:
        - slug
        - name
        - shortDescription
        - fullDescription
        - iconType
        - iconColor
        - windowType
        - status
        - category
        - sortOrder
        - featured

    DivisionUpdate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortDescription:
          type: string
        websiteUrl:
          type: string
          nullable: true
        status:
          type: string
        featured:
          type: boolean
      required:
        - id

    TrackCreate:
      type: object
      properties:
        artistId:
          type: integer
        albumId:
          type: integer
        title:
          type: string
        duration:
          type: integer
        trackNumber:
          type: integer
        previewUrl:
          type: string
          nullable: true
        featured:
          type: boolean
      required:
        - artistId
        - albumId
        - title
        - duration
        - trackNumber
        - featured

    ArtistCreate:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        bio:
          type: string
        genre:
          type: string
        imageUrl:
          type: string
          nullable: true
        featured:
          type: boolean
      required:
        - slug
        - name
        - bio
        - genre
        - featured

    AlbumCreate:
      type: object
      properties:
        artistId:
          type: integer
        title:
          type: string
        coverUrl:
          type: string
          nullable: true
        releaseYear:
          type: integer
        genre:
          type: string
        price:
          type: number
        albumType:
          type: string
          enum: [album, ep, single]
        featured:
          type: boolean
      required:
        - artistId
        - title
        - releaseYear
        - genre
        - price
        - albumType
        - featured

    Artist:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        bio:
          type: string
        genre:
          type: string
        imageUrl:
          type: string
          nullable: true
        featured:
          type: boolean
      required:
        - id
        - slug
        - name
        - bio
        - genre
        - featured

    ArtistDetail:
      allOf:
        - $ref: "#/components/schemas/Artist"
        - type: object
          properties:
            albums:
              type: array
              items:
                $ref: "#/components/schemas/Album"
            topTracks:
              type: array
              items:
                $ref: "#/components/schemas/Track"
          required:
            - albums
            - topTracks

    Album:
      type: object
      properties:
        id:
          type: integer
        artistId:
          type: integer
        artistName:
          type: string
        title:
          type: string
        coverUrl:
          type: string
          nullable: true
        releaseYear:
          type: integer
        genre:
          type: string
        price:
          type: number
        albumType:
          type: string
          enum: [album, ep, single]
        featured:
          type: boolean
        trackCount:
          type: integer
      required:
        - id
        - artistId
        - artistName
        - title
        - releaseYear
        - genre
        - price
        - albumType
        - featured
        - trackCount

    AlbumDetail:
      allOf:
        - $ref: "#/components/schemas/Album"
        - type: object
          properties:
            tracks:
              type: array
              items:
                $ref: "#/components/schemas/Track"
          required:
            - tracks

    Track:
      type: object
      properties:
        id:
          type: integer
        albumId:
          type: integer
        artistId:
          type: integer
        artistName:
          type: string
        albumTitle:
          type: string
        title:
          type: string
        trackNumber:
          type: integer
        duration:
          type: integer
          description: Duration in seconds
        audioUrl:
          type: string
          nullable: true
        previewUrl:
          type: string
          nullable: true
        price:
          type: number
        isExplicit:
          type: boolean
        streamable:
          type: boolean
        downloadable:
          type: boolean
        featured:
          type: boolean
      required:
        - id
        - albumId
        - artistId
        - artistName
        - albumTitle
        - title
        - trackNumber
        - duration
        - price
        - isExplicit
        - streamable
        - downloadable
        - featured

    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        filename:
          type: string
        category:
          type: string
          enum: [corporate, investor, press, legal, division, services]
        content:
          type: string
        fileType:
          type: string
          enum: [txt, doc, pdf, ppt]
        divisionId:
          type: integer
          nullable: true
        public:
          type: boolean
      required:
        - id
        - title
        - filename
        - category
        - content
        - fileType
        - public

    AdminStats:
      type: object
      properties:
        totalDivisions:
          type: integer
        activeDivisions:
          type: integer
        totalTracks:
          type: integer
        totalArtists:
          type: integer
        totalDocuments:
          type: integer
      required:
        - totalDivisions
        - activeDivisions
        - totalTracks
        - totalArtists
        - totalDocuments