# lib/api-zod/src/generated/types/trackCreate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NJAYCO OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrackCreate {
  artistId: number;
  albumId: number;
  title: string;
  duration: number;
  trackNumber: number;
  previewUrl?: string | null;
  featured: boolean;
}